requirements.txt
fastapi~=0.115.12
redis~=6.2.0
httpx~=0.28.1
langchain-google-genai~=2.1.5
langchain-core~=0.3.63
python-dotenv~=1.1.0
uvicorn~=0.34.3
h11~=0.16.0
jupyter~=1.1.1
pip~=21.2.3
attrs~=25.3.0
wheel~=0.45.1
protobuf~=6.31.1
pillow~=11.2.1
tornado~=6.5.1
Jinja2~=3.1.6
filelock~=3.18.0
rsa~=4.9.1
pyasn1~=0.6.1
grpcio~=1.72.1
PyYAML~=6.0.2
typing_extensions~=4.14.0
exceptiongroup~=1.3.0
sniffio~=1.3.1
pytest~=8.4.0
click~=8.2.1
Pygments~=2.19.1
rich~=14.0.0
httpcore~=1.0.9
idna~=3.10
zstandard~=0.23.0
proto-plus~=1.26.1
async-timeout~=4.0.3
requests~=2.32.3
ipython~=8.37.0
aiohttp~=3.12.7
pyasn1_modules~=0.4.2
cachetools~=5.5.2
certifi~=2025.4.26
urllib3~=2.4.0
packaging~=24.2
cffi~=1.17.1
grpcio-status~=1.72.1
orjson~=3.10.18
starlette~=0.46.2
pydantic~=2.11.5
pydantic_core~=2.33.2
websockets~=15.0.1
anyio~=4.9.0
gradio~=5.32.1
gunicorn~=23.0.0
filetype~=1.2.0
Deprecated~=1.2.18
annotated-types~=0.7.0
charset-normalizer~=3.4.2
tenacity~=9.1.2
langsmith~=0.3.44
psutil~=7.0.0
numpy~=1.26.4
pandas~=2.2.3
requests-toolbelt~=1.0.0
langchain~=0.3.25
itsdangerous~=2.2.0
jsonpointer~=3.0.0
jsonpatch~=1.33
langchain-text-splitters~=0.3.8
transformers~=4.52.4
python-multipart~=0.0.20
langchain-community~=0.3.24
pypdf
mysql-connector-python